Refactor functions to be more specific about the data they receive
We are passing some big objects around to our functions when on review, we can see that only a single bit from it is being used. We refactor these so we now only pass the specific data we need. We also remove some instances where we are destructuring variables that we then only use once