Refactor persisting logic
We were checking the length of each data set that we wanted to persist to ensure we had some data. However following through the logic of our previous stages shows that if we have added a transaction to our set of data to be persisted then we will also have added a licence and invoice licence. Therefore, we only need to check the size of the transactions array when determining whether we need to persist data. We also have our function return a boolean which we then use as the service's return value, instead of checking the size of the array then checking it again for the service's return value.