Fix isNewLicence null violation for prev. trans.
We had an issue reported that we were able to replicate locally. It looks like when we try to persist `billing_transactions` 'en masse' using Objection.js it behaves slightly differently than when inserting them in one by one.

With our latest changes to to wait till we have generated _all_ the transactions for a billing period before persisting a new issue has come to light.

```
...
null value in column \"is_new_licence\" of relation \"billing_transactions\" violates not-null constraint"
...
```

After some debugging we found that our generated transactions wre ok. It was any previous transactions we had fetched and were trying to persist as credits. We weren't fetching `isNewLicence` because we were depending on it defaulting to false. It seems we must now fetch the property and ensure it is in the object when trying to persist them en masse.