Fix issue using Array.at() - Node is to old (#242)
It was reported that SROC supplementary bill runs were not running in our `dev` environment. But we could run them locally and our CI seemed fine.

We managed to track down the issue to it not being available on the version of Node we are running in AWS. There we are running v14, locally we are running v16.

We decided to 'look ahead' so that when we do upgrade all will be fine. But clearly, this is an example of that coming to bite us in the 🍑!

So, this change reverts the use of `at()` for a more traditional approach. We'll tackle the disparity in Node versions separately.