Split `ProcessBillingPeriodService` into multiple services (#241)
https://eaflood.atlassian.net/browse/WATER-3996

Our `ProcessBillingPeriodService` is doing a lot by itself! We therefore split it into smaller processes.

We create `SendBillingTransactionsService`, which handles the sending of transactions to the Charging Module, and move the pre-generation of billing data into `PreGenerateBillingDataService`

While refactoring `PreGenerateBillingDataService` we spotted a small bug in the code that pre-generates the billing invoice licences. We aim to create one entry for each combination of billing invoice and licence, with a check to ensure we don't generate it again if it already exists. We should have been checking whether `acc[key]` exists (`key` being the variable that contains the identifying key for that combination of billing invoice and licence), however we were actually checking whether `acc.key` exists (ie. a property of `acc` literally called `key`). This isn't really an issue since all that would happen is the existing entry was being overwritten with no other consequences. However we correct this all the same.