Be consistent about "reissue" terminology
The business refers to what we're doing as "reissuing" an invoice. However for historical reasons the Charging Module refers to it as "rebilling". We previously thought it made sense to talk about "rebilling" when dealing with the CM, and "reissuing" everywhere else. However on reflection, we decided that we would prefer to be consistent within our own codebase. We therefore change our terminology to use "reissue" everywhere.