Refactor to accommodate future service
We will be adding our new service into the existing process flow with the intention that it will return a Boolean to state whether any changes were made, similar to how we currently have a `isBatchPopulated` boolean. We therefore simplify the existing logic to remove all conditionality, and instead put all our results in an array which we then check for any `true` results. This means when we introduce our service it's simply one more result to add to the array with no additional logic required.

We will also be timing the reissuing stage separately from the current processing stage, so we make a small change to the point where we get the start time of the process so we can insert our new service before it.