Fix no credits for REPLACED charge versions (#258)
https://eaflood.atlassian.net/browse/WATER-3996

We introduced an error as part of our second pass at tidying up the SROC supplementary billing code. We moved to generate all possible billing invoices and invoice licences upfront as this vastly simplified the code and meant we could fire a single query that returns many results rather than one query per invoice account ID.

What we got wrong when doing this was a check we had that looked at whether a charge version was `current` (APPROVED) or `superseded` (REPLACED). When we refactored we used it to determine whether to consider a charge version when building our billing data. It was there to decide whether to calculate transaction lines for a charge version.

The refactor meant REPLACED charge versions were being ignored which meant we were never looking back at previous transactions (which is the basis for generating credits).

This change fixes the issue and moves the check to where it should be.