Break up `GenerateBillingTransactionsService` into separate services
https://eaflood.atlassian.net/browse/WATER-4036

Currently `GenerateBillingTransactionsService` does the work of calculating the number of billable days, determining if a compensation transaction is required, then converting provided data into the format required by a transaction record.

Doing all this in one service means we can't reuse the data conversion code elsewhere (ie. when creating transactions for reissuing).

This PR aims to break `GenerateBillingTransactionsService` into two separate services: one that calculates the number of billable days and determines if a compensation transaction is required, and one that converts data into the required format. This second service can then be called from elsewhere.