SROC Supp flag not cleared for 'non-chargeable' licence
https://eaflood.atlassian.net/browse/WATER-4041

It has been found in QA that when a licence has no other SROC charge versions and is made 'non-chargeable', this correctly sets the SROC supplementary billing flag on the licence. However, the problem is that when a bill run is created that includes this licence, the SROC supplementary billing flag is not un-set for it.

This issue is being caused because when a licence is made “non-chargeable” it creates a charge version without any charge elements, but also without an `invoice_account_id`.

Our `fetch-charge-versions` service only fetches charge versions which have an `invoice_account_id`. Which in effect means that if a licence only has the “non-chargeable” charge version associated with it. That licence will never be looked at since the licence ID is extracted from the results of the fetch-charge-versions service. As the licence is never looked at by the SROC supplementary service the flag is never un-set.

This PR fixes that issue by initially also fetching the charge versions that have no `invoice_account_id`, rather than filtering them out in the query as was done previously.

The `licence_id`s are then extracted from the results of this query before finally stripping out the charge versions that have no `invoice_account_id`.