Remove unneeded "get invoices"
If getting invoices fails, we silently swallow the error and return `false`, as we can continue with the rest of the billing process. While setting up the initial tests we were also getting invoices outside of the try/catch to ensure any errors in this stage would show up in our terminal. We now have enough unit tests in place that we no longer need this so we remove it. We also add a TODO to consider adding logging of errors here