Refactor to remove top-level try/catch
The first thing we do is fetch the invoices to be reissued from the db, and if there are none then we return `false` so we can continue with processing the billing batch. If fetching invoices throws an error for some reason then we are okay to simply log the error and return `false` so we can continue, since no changes have yet been made (either to the db or to the Charging Module data) that require the billing batch to be marked as errored.

This was being done in a top-level try/catch, which caused some clutter and slightly hid what was going on. We move this down into the `_fetchInvoicesToReissue` function so at the top level we simply handle it as "if there are no invoices to reissue, return false".