Better handle errors in Notifiers
We spotted this issue when [working in another service](https://github.com/DEFRA/water-abstraction-import/pull/661) where we have copied the `app/lib/base-notifier.lib.js` and `app/lib/global-notifier.lib.js` across. At the points we wanted to log an error we were calling `global.GlobalNotifier.omfg('import.charge-versions: errored', error)`. When we got an error though, this is what we saw.

In the logs it would just be

```text
0|import  | [15:44:40.837] ERROR (8616):
0|import  |     message: "import.charge-versions: errored"
```

Errbit would get the notification but if you tried to view it you'd get an error. Looking into this last one it was to do with how the Airbrake package was handling an instance of [Error](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/Error) being passed as the `session:` property in the call to `notify()`. Simple errors were fine but custom ones appeared to break Errbit.

Simply, we were treating instances of `Error` as if they were just a POJO when they are not.

```javascript
const myError = new Error('boom')

const myObject = {
  ...myError
}

console.log(myObject) // {}
```

For Airbrake and Errbit uncaught errors were coming through and displaying fine. The key difference is `app/plugins/airbrake.plugin.js` when calling `notify()` passes in the error instance as the `error:` property.

This change updates our notifier modules to better calls to `omfg()` that pass in an `Error` instance. In the logs we will get

```text
0|import  | [16:42:39.829] ERROR (8786):
0|import  |     message: "import.charge-versions: errored"
0|import  |     error: {
0|import  |       "stack":
0|import  |           error: null value in column "start_date" violates not-null constraint
0|import  |               at Parser.parseErrorMessage (/home/repos/water-abstraction-import/node_modules/pg-protocol/dist/parser.js:287:98)
0|import  |               at Parser.handlePacket (/home/repos/water-abstraction-import/node_modules/pg-protocol/dist/parser.js:126:29)
0|import  |               at Parser.parse (/home/repos/water-abstraction-import/node_modules/pg-protocol/dist/parser.js:39:38)
0|import  |               at Socket.<anonymous> (/home/repos/water-abstraction-import/node_modules/pg-protocol/dist/index.js:11:42)
0|import  |               at Socket.emit (node:events:513:28)
0|import  |               at Socket.emit (node:domain:489:12)
0|import  |               at addChunk (node:internal/streams/readable:315:12)
0|import  |               at readableAddChunk (node:internal/streams/readable:289:9)
0|import  |               at Socket.Readable.push (node:internal/streams/readable:228:10)
0|import  |               at TCP.onStreamRead (node:internal/stream_base_commons:190:23)
0|import  |       "message": "null value in column \"start_date\" violates not-null constraint",
0|import  |       "length": 307,
0|import  |       "name": "error",
0|import  |       "severity": "ERROR",
0|import  |       "code": "23502",
0|import  |       "detail": "Failing row contains (4:10024704:licenceStart-licenceEnd, 1, null, 2018-12-31, current, t, 2023-06-16 16:42:39.787071+00, null).",
0|import  |       "schema": "water_import",
0|import  |       "table": "charge_versions_metadata",
0|import  |       "column": "start_date",
0|import  |       "file": "execMain.c",
0|import  |       "line": "1948",
0|import  |       "routine": "ExecConstraints"
0|import  |     }
```

In Errbit all errors will be visible. Also, it will be able to determine the error class and therefore better group notifications.