Add new serializeError() method
hapi-pino wants a POJO in order to output something to the logs. But as demonstrated we can't just use `{ ...myError }` to do that.

We used the responses to [Is it not possible to stringify an Error using JSON.stringify?](https://stackoverflow.com/q/18391212) to implement a solution that uses `Object.getOwnPropertyNames()` to identify the properties directly on the error instance. So, `message:` and `stack:` will be there as all errors inherit these. But if a custom error adds additional properties then we'll get those too.

That comes back as an array, which we iterate to extract the values off the error and create a new POJO from.