Revert last change and implement default format
First, we'd forgotten that `omg()` calls `.info()` and `omfg()` calls `.error()` on the logger.

So, we went back to the docs to double check there wasn't something we were missing that is different about these methods.

The docs confirmed the signature is the same. What we'd overlooked (either originally or they've changed on us!) is that the log methods expect any data value to be passed through first. When you have a value to log, the second arg should be the message. We were formatting a single object with a `message:` property.

The first arg is referred to as the [mergingObject](https://github.com/pinojs/pino/blob/master/docs/api.md#mergingobject-object)

> Each enumerable key and value of the `mergingObject` is copied into the JSON log line.

We could have made our current implementation work if we amended the `message:` property to be `msg:`. Then pino would know to use that as the message for the log entry.

But if the `mergingObject` is an error and no `message` arg is provided then pino will use the errors `message:` property instead. We want to be able to log an error correctly _and_ still have our own message. For this you must provide both `mergingObject` and `message` args to pino. So, it makes sense at this stage to always be passing both args to pino and stick with that convention.

The second key change is we're scraping the idea of forcing notifiers that extend the base to implement their own formatters. Experience has shown we're just copying the same format logic in them so we may as well leave that in the base as a default implementation.

This change updates BaseNotifier with a default implementation for `_formatLogPacket()`. We then update the tests to cover both changes.