Ignore all previous changes!
So, we started thinking we had this licked. But that refresher of the Pino docs and the move to have the BaseNotifier implement the default formatters open up a different perspective.

That led to what is implemented here. Essentially, when logging errors we _always_ log an error, even if we have to generate one. For notifications we always send an error to Errbit. Again, it helps keep things consistent and supports Errbit with its grouping.

We also tweaked what to do when things go wrong. If Airbrake fails we no longer use `_formatLogPacket()`. We just call the `logger.error()` directly to avoid issues.

Having tested with real errors we know this delivers the results we want now. Hopefully, the unit tests express the behaviour we want.