Update RequestNotifier
This is currently the one notifier we have that was overriding what BaseNotifier was doing in the _format() methods (beyond the copy & paste default implementation used by the others).

Now BaseNotifier holds our default implementation, we simply want to inject our behaviour in the process. The RequestNotifier needs to add the request ID to the data `mergeObject` pino will log and the session object Airbrake sends.

We now override the base methods to create a new decorated object based on what is passed in. We then forward that to the BaseNotifier methods using [super](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/super).