Fix supp. billing not crediting old accounts
https://eaflood.atlassian.net/browse/WATER-4044

Assume the following scenario. A charge version is added to a licence starting on 1 April 2022 against invoice account **INVAC001**. The licence is flagged for supplementary billing which is run during the 2023-2024 billing period.

This will will generate 2 bills for **INVAC001**; one in 2024 and one in 2023 ([multi-year billing!](https://github.com/DEFRA/water-abstraction-system/pull/228))

Then another charge version is added starting on 1 June 2022 against invoice account **INVAC002**. Again the licence is flagged for supplementary billing which is run during the 2023-2024 billing period.

What should happen is

- 2024
  - **INVAC001** receives a whole year credit
  - **INVAC002** receives a whole year debit
- 2023
  - **INVAC001** receives a part year credit and debit
  - **INVAC002** receives a part year debit

So, 4 bills in total which result in a £0 bill run.

What is happening is no credit is generated for **INVAC001** in 2024. The reason is `FetchChargeVersionsService` was written to exclude charge versions which start and finish outside the billing period being processed. But that exclusion means we never check them for previous transactions, which in this scenario would result in **INVAC001's** debit being found so a credit generated.

This change resolves the issue by removing the clause in our query which excludes charge versions that end before the billing period start date. To support this we also need to amend the `_periodIsInvalid()` in `DetermineChargePeriodService`.

> Note - removal of this clause means we might pick up charge versions that don't need to be processed at all. We did manage to implement an Objection.js query that used a [UNION](https://www.w3schools.com/sql/sql_union.asp) to prevent this. But it made `FetchChargeVersionsService` vastly more complex purely for a negligible performance bump.