Add default dev/test user password
We know we want to seed the dev/test users as our first 'test' of seeding. We want to take the opportunity to move away from hard coding the password we use in the source.

So, in preparation for this we add a new env var, something we do on all our previous services.