Add first seed file for users
First thing to get your head around is that unlike migrations Knex always runs _all_ seed files. So, it's important to write them in the knowledge the data they are seeding might already exist.

For users, our initial seed example, we intend to only ever seed them once. So, we have implemented a method that first checks if a user already exists. If it doesn't we create the new user.

This then applies to the user groups, which determines the permissions a user has in the service.

To be able to log in with these users the password we generate must be recognised by the legacy code. So, we're forced to use the same mechanism currently used in [water-abstraction-tactical-idm](https://github.com/DEFRA/water-abstraction-tactical-idm) even if the package now appears to be no longer maintained.