Add endpoint to trigger seed process
Normally you would run the Knex seed process from the command line as part of your deployment. But we intend to use this mechanism within our acceptance tests. This means we need a method to programmatically trigger the process.

This adds a new endpoint, which with some help from Stackoverflow allows us to do that!