Refactor `isBatchPopulated` logic
We originally had a `isBatchPopulated` flag which we set `false` at the start of processing, then updated to `true` if the call to `ProcessBillingPeriodService` came back as `true` at any point. To avoid duplicating logic when we introduce reissuing, we refactor this to add the result to an array and then set `isBatchPopulated` based on whether at least one result in the array is truthy.