Fix licence ends before billing period
https://eaflood.atlassian.net/browse/WATER-4044

We recently implemented a [fix to stop calculating negative billable days for out of period charge versions](https://github.com/DEFRA/water-abstraction-system/pull/286).

Each billable period (financial year) we process we have to look at all charge versions that start before the period ends. This is to catch those which were previously debited in the period but recent changes now cause them to end before the billable period.

We've since spotted another scenario that causes a charge period to be returned where the start date is after the end date. A charge version that starts before the billable period but is linked to a licence that ended before the billable period results in an invalid charge period

- charge version start date 2022-04-01
- licence revoked date 2023-01-11
- billable period start date 2023-04-01
- billable period end date 2024-03-31

Our current `_periodIsIncompatible()` check would determine the charge version has a compatible period. So, the service would then grab the latest start date; billable period start date. Next it would grab the earliest end date; licence revoked date.

The result is a charge period that results in negative billable days

```javascript
  return {
    startDate: new Date('2023-04-01'),
    endDate: new Date('2023-01-11')
  }
```

This change moves the `_periodIsIncompatible()` check to after we've determined the start and end dates. This should mean no matter the reason for determining a start date after the end date, the service will see it is incompatible and return the null result needed to stop `ProcessBillingPeriodService` trying to calculate billable days.