Move test year up by 1
When we originally wrote this service we were only having to deal with the current year, which also happened to be the first year of SROC (2022-04-01 to 2023-03-31).

We are now trying to add scenarios that cover charge versions that start before the 'current' billable period. We can use 2021 to 2022, but that seems wrong when we know it's not actually a valid SROC billing period.

So, before we start adding our test cases we update the existing tests by bumping all the years forward by 1. We can then add our scenarios for stuff that starts and ends before the billing period of 2023-2024 and _still_ have everything be inside SROC.