Add tests for problem scenario
It was a revoked date that was the root of our problem scenario but the same could happen for expired and lapsed licences.

So, we've extended the existing tests to cover what happens if the licences ends before the billing period. They all currently fail. Next we just need to implement the fix.