Update service to receive period
We realised the service was working out the billing period using the `financialYearEnding` value passed in. Where does it get this value? Why, from the `ProcessBillingPeriodService` passing in `billingPeriod.endDate.getFullYear()` 🤦

So, we take this chance to update the service to just expect the billing period to have been worked out and passed in.