Fix broken DetermineMinimumChargeService
After making the change to have `DetermineChargePeriodService` just accept a `billingPeriod` param we'd broken supplementary billing!

A quick check found that `DetermineMinimumChargeService` was calling `DetermineChargePeriodService` to work out the charge period. It was still using the financial year end arg. So, initial though was we'll need to get the billing period to it as well.

But then we realised that `DetermineMinimumChargeService` is only called by `ProcessBillingPeriodService` and this happens straight after it's called `DetermineChargePeriodService`. We already have the `chargePeriod`. There is no need to calculate it again.