Add new scenario
We realised there is another possible scenario not currently catered for.

What if the charge version being processed starts after the expired date of the licence, and both those dates are within the billing period?

The service is going to use the charge version's start date as that is the latest but the licence expired date as that is the earliest. This results in another charge period where the start date is after the end date.