Rework generation of transactions
The Charging Module does the work of determining whether transactions are credits or debits. We can therefore refactor the service so that instead of working this out for ourselves based on whether we're looking at the cancelling invoice or the reissuing invoice, we can simply use the transaction's credit flag.