Implement SROC invoice reissuing (#256)
https://eaflood.atlassian.net/browse/WATER-4036

We have [created the new services we need](https://github.com/DEFRA/water-abstraction-system/pull/255) to send the appropriate requests to the Charging Module. We now implement invoice reissuing for SROC.

This is done with `ReissueInvoicesService`, which calls `FetchInvoicesToBeReissuedService` to get the invoices for reissuing then passes each one to `ReissueInvoiceService` (note singular `Invoice`) which handles the actual reissuing of the invoice. The resulting invoices, invoice licences and transactions are persisted, and a `true` value returned to indicate that db changes have been made.

The intent is that this service will be called from `ProcessBillingBatchService`, and its return value used to set the `isBatchPopulated` boolean. This PR initially made some changes to `ProcessBillingBatchService` to accommodate this; however due to a merge conflict (and the fact that the changes don't really belong in this PR) the changes in the [`Refactor to accommodate future service`](https://github.com/DEFRA/water-abstraction-system/pull/256/commits/b251f8a4ba2bd77c74d55021badf4c54c02633da) commit have been reverted. They will be reinstated in the future PR which adds reissuing to the process billing batch process.