Add initial seeding support using users (#285)
https://eaflood.atlassian.net/browse/WATER-3981

The [water-abstraction-service](https://gitub.com/DEFRA/water-abstraction-service) has a mechanism for seeding data for testing that relies on loading in YAML fixture files, sort of. Some content comes from these but other stuff is hard-coded into the 'loaders'.

Also, because you often need to add the ID of a record into something else your seeding it supports a system of 'named references' so the fixture-loader knows how to link things. Again though, it depends if the specific fixture-loader has been written to support this.

What we've been left with is an extremely complex, brittle seeding mechanism that is also very slow. If only there was something simpler that existed to do this!?

There is, and we're already using it. [Knex](https://knexjs.org/guide/migrations.html#seed-files) supports seeding using the same syntax we already use elsewhere in the app. It's something we've used previously on the [charging-module-api](https://github.com/DEFRA/sroc-charging-module-api).

We think this is a better way to go and means we can start to move away from the hand-cranked version in **water-abstraction-service**.

This change gets the ball rolling by creating a seed for users, and hooking it into the project so it can be triggered as a package.json script or from an endpoint (needed to support [water-abstraction-acceptance-tests](https://github.com/DEFRA/water-abstraction-acceptance-tests)).