Revert time logging
Some early feedback was that we don't care so much about the time taken to reissue invoices -- we're more concerned with how long the overall process takes. So we revert the time logging to how things were prior to this PR -- that is, we log the overall process time, and we do this within the `go` function to ensure it covers the entire process (ie. reissuing as well as processing the billing periods)