Move supplementary-billing to new billing folder
https://eaflood.atlassian.net/browse/WATER-4057

We are on the cusp of adding our second bill run type; two-part tariff. Our experience of implementing the first, supplementary billing has shown that we're going to need a lot of services! 😁

So, ahead of starting to implement we are doing some housekeeping in our services folder. The first of these was to [Move db-export services to data folder](https://github.com/DEFRA/water-abstraction-system/pull/299). The second of these is to move the `services/supplementary-billing/` to `services/billing/supplementary/`.