Add new check TwoPartService
This will serve as a wrapper for calling the services we'll start building to support 2PT bill runs.

We know just like supplementary region is the primary filter for which licences to include in the bill run. So, to get us started with something we build it with the expectation that we need to fetch it.

Side note - looks like we've found our first generic billing service!