Update check endpoint
First thing we need to amend is the type of endpoint. The point is to aid with testing that we're returning the right results. Nothing is being created so we should be using GET not POST!

Beyond that we hook the controller into our new check service and update the tests accordingly.