Add back in our check endpoint (#302)
https://eaflood.atlassian.net/browse/WATER-4038

Way back when we first started the work on SROC Supplementary billing we [started with a `/test` endpoint](https://github.com/DEFRA/water-abstraction-system/pull/13). We used this as a means of triggering our billing engine and exposing the results, until such time as we were ready to integrate it into the process for real.

We're about ready to start developing the next bill run type, which is two-part tariff (2PT) and we want to do the same again. Over time we learnt things were much less confusing if the endpoint was actually `/check` (testing `TestService` in the `test/services/test` folder - need we say more!)

So, this change adds the endpoint back into the repo. We can then start iterating our 2PT engine, verifying the results along the way.