Make create bill run validator generic
We update the validator to accept both supplementary and 2PT arguments (based on what the UI will fire at us).

In the [charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) we used the convention of keeping any static list of values in a central place.

This seemed an apt time to bring in that concept to this project and make use of it in the validator.